moved Spinner.jsx into components/ui and added use client
